# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ASTAR Task",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/items": {
            "get": {
                "tags": [
                    "Item"
                ],
                "summary": "show all Items",
                "description": "show all Items Here",
                "operationId": "showallitems",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "perPage": {
                                        "type": "integer"
                                    },
                                    "onlyRequested": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "multipart/form-data": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Item"
                ],
                "summary": "create items ",
                "description": "create items Here",
                "operationId": "createitems",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "partition_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "partition_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item created Successfully",
                        "content": {
                            "multipart/form-data": {
                                "schema": {
                                    "properties": {
                                        "item": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/items/{item}": {
            "get": {
                "tags": [
                    "Item"
                ],
                "summary": "get item By Id",
                "description": "get item By Id Here",
                "operationId": "getitemById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get item By Id",
                        "content": {
                            "multipart/form-data": {
                                "schema": {
                                    "properties": {
                                        "item": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Item"
                ],
                "summary": "update items ",
                "description": "update items Here",
                "operationId": "updateitems",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "partition_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item updated Successfully",
                        "content": {
                            "multipart/form-data": {
                                "schema": {
                                    "properties": {
                                        "item": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/items/{id}": {
            "delete": {
                "tags": [
                    "Item"
                ],
                "summary": "Delete item",
                "description": "delete item here",
                "operationId": "delete-item",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "item deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/items/restore/{id}": {
            "post": {
                "tags": [
                    "Item"
                ],
                "summary": "restore item By Id",
                "description": "restore item By Id Here",
                "operationId": "restoreitemById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "restore item By Id",
                        "content": {
                            "multipart/form-data": {
                                "schema": {
                                    "properties": {
                                        "item": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}